+++ packages/atomic/cypress/integration/component-error-selectors.ts
@@ -1,0 +1,3 @@
+export const ComponentErrorSelectors = {
+  component: 'atomic-component-error',
+};