+++ packages/atomic/cypress/integration/facets/facet.cypress.ts
@@ -6,6 +6,7 @@ import {doSortAlphanumeric} from '../../utils/componentUtils';
 
 import {
   validateFacetComponentLoaded,
+  validateFacetComponentVisible,
   validateFacetNumberofValueEqual,
   assertBasicFacetFunctionality,
   assertSortCriteria,
@@ -322,8 +323,12 @@ describe('Facet with invalid options', () => {
       shouldRenderErrorComponent(FacetSelectors.facetStandard);
     });
 
-    it.skip('Should not render when field returns no result', () => {
+    it('Should not render when field returns no result', () => {
       setupFacet('author2', facetProp.label);
+      validateFacetComponentVisible(
+        facetProp.label,
+        FacetSelectors.facetStandard
+      );
     });
   });