+++ packages/atomic/cypress/integration/result-list.cypress.ts
@@ -1,19 +1,66 @@
+import {setUpPage} from '../utils/setupComponent';
+import {createAliasNavigation, PagerSelectors} from './pager-selectors';
+import {
+  ResultListSelectors,
+  resultListComponent,
+} from './result-list-selectors';
+
 describe('Result List Component', () => {
-  it.skip('should load');
+  function getFirstResult() {
+    return cy
+      .get(ResultListSelectors.component)
+      .find(ResultListSelectors.result)
+      .first()
+      .shadow();
+  }
+
+  it('should load', () => {
+    setUpPage(resultListComponent());
+    cy.get(ResultListSelectors.component).should('be.visible');
+  });
 
   describe('when no first search has yet been executed', () => {
-    it.skip('should render a "atomic-result-list-placeholder" component');
+    beforeEach(() => {
+      setUpPage(resultListComponent(), false);
+    });
+
+    it('should render a placeholder component', () => {
+      cy.get(ResultListSelectors.component)
+        .find(ResultListSelectors.placeholder)
+        .should('be.visible');
+    });
   });
 
   describe('when an initial search is executed', () => {
-    it.skip('should render the correct number of results');
+    it('should render the correct number of results', () => {
+      setUpPage(resultListComponent());
+      cy.get(ResultListSelectors.component)
+        .find(ResultListSelectors.result)
+        .should('have.length', 10);
+    });
   });
 
   describe('when multiple searches are executed', () => {
-    it.skip('should update the results');
-  });
+    it('should update the results', () => {
+      let firstResultHtml: string;
+      setUpPage(
+        `${resultListComponent()}<${PagerSelectors.pager}></${
+          PagerSelectors.pager
+        }>`
+      );
+      createAliasNavigation();
 
-  describe('when a result template contains an error', () => {
-    it.skip('should render an "atomic-component-error" component');
+      getFirstResult().then((element) => {
+        firstResultHtml = element[0].innerHTML;
+      });
+
+      cy.get('@nextButton').click();
+      cy.wait(500);
+
+      getFirstResult().should((element) => {
+        const secondResultHtml = element[0].innerHTML;
+        expect(secondResultHtml).not.to.equal(firstResultHtml);
+      });
+    });
   });
 });