+++ packages/atomic/cypress/integration/result-templates/result-template-selectors.ts
@@ -1,0 +1,6 @@
+export const ResultTemplateSelectors = {
+  component: 'atomic-result-template',
+};
+
+export const resultTemplateComponent = (slot = '', attributes = '') =>
+  `<${ResultTemplateSelectors.component} ${attributes}>${slot}</${ResultTemplateSelectors.component}>`;