+++ packages/atomic/cypress/utils/setupComponent.ts
@@ -1,11 +1,22 @@
-export function injectComponent(componentInCode: string) {
-  cy.get('atomic-search-interface').should('exist');
-  cy.document().then((document: any) => {
-    document.querySelector(
-      'atomic-search-interface'
-    ).innerHTML = componentInCode;
+const searchInterfaceTag = 'atomic-search-interface';
+export function injectComponent(
+  componentHtml: string,
+  executeFirstSearch = true
+) {
+  cy.document().then(async (document) => {
+    document.body.innerHTML = `<${searchInterfaceTag}>${componentHtml}</${searchInterfaceTag}>`;
+    const searchInterface: any = document.querySelector(searchInterfaceTag);
+    await searchInterface.initialize({
+      accessToken: 'xx564559b1-0045-48e1-953c-3addd1ee4457',
+      organizationId: 'searchuisamples',
+    });
+    executeFirstSearch && searchInterface.executeFirstSearch();
   });
 }
+
+export const searchEndpoint =
+  'https://platform.cloud.coveo.com/rest/search/v2?organizationId=searchuisamples';
+
 export function setupIntercept() {
   cy.intercept({
     method: 'POST',
@@ -19,18 +30,19 @@ export function setupIntercept() {
 
   cy.intercept({
     method: 'POST',
-    url:
-      'https://platform.cloud.coveo.com/rest/search/v2?organizationId=searchuisamples',
+    url: searchEndpoint,
   }).as('coveoSearch');
 }
 
-export function setUpPage(htmlCode: string) {
+// TODO: rename to setupPage (typo)
+// TODO: add options object for arguments (with urlHash, wait options)
+export function setUpPage(htmlCode: string, executeFirstSearch = true) {
   setupIntercept();
-  // Setup page with new component
   cy.visit('http://localhost:3333/pages/test.html');
   cy.injectAxe();
-  injectComponent(htmlCode);
-  cy.wait(500);
+  injectComponent(htmlCode, executeFirstSearch);
+  // TODO: when executeFirstSearch = true, waiting for @coveoSearch would be less flaky
+  cy.wait(300);
 }
 
 export function shouldRenderErrorComponent(selector: string) {