+++ packages/atomic/src/components/result-template-components/result-template-decorators.ts
@@ -2,8 +2,8 @@ import {ComponentInterface, getElement} from '@stencil/core';
 
 export function ResultContext() {
   return (component: ComponentInterface, resultVariable: string) => {
-    const {render} = component;
-    component.render = function () {
+    const {componentWillLoad} = component;
+    component.componentWillLoad = function () {
       const element = getElement(this);
       const parentResultComponent = (element.getRootNode() as ShadowRoot)
         .host as HTMLAtomicResultElement;
@@ -11,13 +11,15 @@ export function ResultContext() {
         !parentResultComponent ||
         parentResultComponent.nodeName !== 'ATOMIC-RESULT'
       ) {
-        throw new Error(
+        console.error(
           `The "${element.nodeName.toLowerCase()}" component has to be used inside a template element 
           https://developer.mozilla.org/en-US/docs/Web/HTML/Element/template`
         );
+        element.remove();
+        return undefined;
       }
       component[resultVariable] = parentResultComponent.result;
-      return render && render.call(this);
+      return componentWillLoad && componentWillLoad.call(this);
     };
   };
 }