+++ packages/samples/headless-react/src/components/result-list/result-list.class.tsx
@@ -7,6 +7,7 @@ import {
 } from '@coveo/headless';
 import {ResultLink} from './result-link';
 import {AppContext} from '../../context/engine';
+import {Quickview} from '../quickview/quickview.class';
 
 export class ResultList extends Component<{}, ResultListState> {
   static contextType = AppContext;
@@ -45,11 +46,14 @@ export class ResultList extends Component<{}, ResultListState> {
           {this.state.results.map((result) => (
             <li key={result.uniqueId}>
               <article>
-                <h2>
+                <h3>
                   {/* Make sure to log analytics when the result link is clicked. */}
                   <ResultLink result={result}>{result.title}</ResultLink>
-                </h2>
+                </h3>
                 <p>{result.excerpt}</p>
+                <div>
+                  <Quickview result={result} />
+                </div>
               </article>
             </li>
           ))}